Improved tests readability, fixed recall and precision. Added new metrics in objective.
